backend/pyproject.toml
[project]
name = "style-transfer-backend"
version = "0.1.0"
description = "3D Islamic Architecture Style Transfer Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "aiosqlite>=0.20.0",
    "python-multipart>=0.0.9",
    "trimesh>=4.4.0",
    "pyrender>=0.1.45",
    "numpy>=1.26.0",
    "Pillow>=10.0.0",
    "httpx>=0.27.0",
    "celery[redis]>=5.4.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]
gpu = ["torch", "diffusers", "transformers"]

[tool.ruff]
line-length = 100
